Add files via upload
Again, I do not know where to get a reference at this point to Dereks exact file structure, so what I decided to do what to attach this which I think is more built out probably than where we would have been in this point in time in class. This is a "starter" file structure for React + Redux with blueprints that I am intending to use for some of my portfolio pieces after reviewing many file structures, starters, etc that has the blueprinting mentioned in the GC assignment.